Improve `.gitignore` check performance
It turns out that the `.gitignore` check in `OmniParser` is very slow, because a new `TreeWalk` is set up for each file to be checked.

This commit changes the `acceptedPaths()` implementation so that it uses a separate implementation when processing Git repos, as a JGit `TreeWalk` is essentially the same thing as what `Files.walkFileTree()` does.